Test 'illigal constructor'
